Huge refactoring, changed all floats to doubles in Aim code.  It's now giving exactly the same results as the Aim code.  Plus, I completely changed how the data flow works between the modules.  Instead of having a Fanout, each of the Aim Marsystems are responsible for outputting the centre frequencies and strobes as well as the signals.  It makes the dataflow between Aim MarSystems much easier, but to integrate them with other MarSystems, we'll need to have a better Selector, which I will write eventually.
git-svn-id: svn+ssh://svn.code.sf.net/p/marsyas/code/trunk@4039 0e98fd9d-079e-457f-9c5c-d482923f975d

+++ src/marsyas/AimHCL.h
@@ -69,12 +69,12 @@ private:
   MarControlPtr ctrl_lowpass_order_;  // Order of Lowpass Filter
 
   // int channel_count_;  // Internal record of the number of channels in the input
-  float time_constant_;  // Time constant corresponsing to the lowpass filter cutoff freqency
+  double time_constant_;  // Time constant corresponsing to the lowpass filter cutoff freqency
 
   // Lowpass filter state variables
-  float xn_;
-  float yn_;
-  std::vector<std::vector<float> > yns_;
+  double xn_;
+  double yn_;
+  std::vector<std::vector<double> > yns_;
 
 public:
   AimHCL(std::string name);
